Endure extreme pain to successfully query MongoDB
Turns out when there's a syntax error of a certain nature nothing happens, no errors, no warnings nothing. It makes it appear our server call hung forever but really the javascript in the script was invalid. We had to detect and correct that before things would work.

Another element of extreme pain was the wide variety of totally different MongoDB client connection directions that were often out of date, incorrect and invalid.