Start Node taking stage-0 preset into account
I can run the server without any preset though. This definitely means my understanding of what is going on is flawed.

Why does Node even need to know? Doesn't webpack need to know this and to use it in order to produce vanilla JS for Node? Webpack definitely croaks the second we use something it isn't preset for that's for sure.